backend/requirements.txt
fastapi
httpx
matplotlib
numpy
pandas
QuantLib
scikit-learn
xgboost
yfinance
alpaca-py
websockets
uvicorn[standard]
pydantic
